ImBox

+++ Sandboxie/core/drv/verify.h
@@ -82,10 +82,10 @@ enum ECertLevel {
     eCertMaxLevel       = 0b111,
 };
 
-#define CERT_IS_TYPE(cert,t)        ((cert.type & 0b11100) == t)
+#define CERT_IS_TYPE(cert,t)        ((cert.type & 0b11100) == (unsigned long)(t))
 #define CERT_IS_SUBSCRIPTION(cert)  (CERT_IS_TYPE(cert, eCertBusiness) || CERT_IS_TYPE(cert, eCertSubscription) || cert.type == eCertEntryPatreon || CERT_IS_TYPE(cert, eCertEvaluation))
 #define CERT_IS_INSIDER(cert)		(CERT_IS_TYPE(cert, eCertEternal) || cert.type == eCertGreatPatreon)
-#define CERT_IS_LEVEL(cert,l)       (cert.active && cert.level >= l)
+#define CERT_IS_LEVEL(cert,l)       (cert.active && cert.level >= (unsigned long)(l))
 
 #ifdef KERNEL_MODE
 extern SCertInfo Verify_CertInfo;
